+++ app/frontend/components/crystals/show/crs_view/action_btns.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {EditBtn} from 'components/crystals/show/crs_view/actionbox/edit_btn'
-import {DeleteBtn} from 'components/crystals/show/crs_view/actionbox/delete_btn'
+import {EditBtn} from 'components/crystals/show/crs_view/action_btns/edit_btn'
+import {DeleteBtn} from 'components/crystals/show/crs_view/action_btns/delete_btn'
 
-export const Actionbox = ({isSelf, crsVals}) => {
+export const ActionBtns = ({isSelf, crsVals}) => {
   /* crystal 所有者は編集・削除可能 */
-  const actionBtns = (selfBool) => {
-    if (!selfBool) return null
+  const actionBtns = () => {
+    if (!isSelf) return null
     return (
       <div className='crsAction'>
         <EditBtn crsVals={crsVals} />
@@ -16,10 +16,10 @@ export const Actionbox = ({isSelf, crsVals}) => {
     )
   }
 
-  return actionBtns(isSelf)
+  return actionBtns()
 }
 
-Actionbox.propTypes = {
+ActionBtns.propTypes = {
   isSelf: PropTypes.bool,
   crsVals: PropTypes.object
 }
