+++ app/frontend/components/fragments/draftjs/frg_view/action_btns.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {EditBtn} from 'components/fragments/draftjs/frg_view/actionbox/edit_btn'
-import {DeleteBtn} from 'components/fragments/draftjs/frg_view/actionbox/delete_btn'
+import {EditBtn} from 'components/fragments/draftjs/frg_view/action_btns/edit_btn'
+import {DeleteBtn} from 'components/fragments/draftjs/frg_view/action_btns/delete_btn'
 
-export const Actionbox = ({isSelf, frgVals}) => {
+export const ActionBtns = ({isSelf, frgVals}) => {
   /* fragment 所有者は編集可能 */
   const buttons = () => {
     if (!isSelf) return null
@@ -19,7 +19,7 @@ export const Actionbox = ({isSelf, frgVals}) => {
   return buttons()
 }
 
-Actionbox.propTypes = {
+ActionBtns.propTypes = {
   isSelf: PropTypes.bool,
   frgVals: PropTypes.object
 }
