+++ app/frontend/components/fragments/draftjs/frg_view/action_btns/delete_btn.jsx
